abstract a new _SummarySession from the rest of `TrainingSession`
